+++ pom.xml
@@ -66,7 +66,7 @@ http://maven.apache.org/maven-v4_0_0.xsd">
           <url>http://localhost:8080/manager/text</url>
           <username>admin</username>
           <password>admin</password>
-          <path>/app</path>
+          <path>/application</path>
         </configuration>
       </plugin>
       <plugin>
