--- src/main/java/com/geekbrains/AppConfig.java
@@ -1,47 +1,0 @@
-package com.geekbrains;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import org.thymeleaf.spring5.SpringTemplateEngine;
-import org.thymeleaf.spring5.templateresolver.SpringResourceTemplateResolver;
-import org.thymeleaf.spring5.view.ThymeleafViewResolver;
-
-@EnableWebMvc
-@Configuration
-@ComponentScan("com.geekbrains")
-public class AppConfig implements WebMvcConfigurer {
-
-  @Override
-  public void addResourceHandlers(ResourceHandlerRegistry registry) {
-    registry.addResourceHandler("/resources/**").addResourceLocations("/resources/"
-    );
-  }
-
-  @Bean
-  public SpringResourceTemplateResolver templateResolver() {
-    SpringResourceTemplateResolver templateResolver = new SpringResourceTemplateResolver();
-    templateResolver.setPrefix("/WEB-INF/templates/");
-    templateResolver.setSuffix(".html");
-    return templateResolver;
-  }
-
-  @Bean
-  public SpringTemplateEngine templateEngine() {
-    SpringTemplateEngine templateEngine = new SpringTemplateEngine();
-    templateEngine.setTemplateResolver(templateResolver());
-    return templateEngine;
-  }
-
-  @Bean
-  public ThymeleafViewResolver thymeleafViewResolver() {
-    ThymeleafViewResolver thymeleafViewResolver = new ThymeleafViewResolver();
-    thymeleafViewResolver.setTemplateEngine(templateEngine());
-    thymeleafViewResolver.setCharacterEncoding("UTF-8");
-    return thymeleafViewResolver;
-  }
-}
-
