--- src/main/java/com/geekbrains/Author.java
@@ -1,22 +1,0 @@
-package com.geekbrains;
-
-public class Author {
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    private int id;
-    private String firstName;
-}
