+++ src/main/java/com/geekbrains/model/Author.java
@@ -1,0 +1,22 @@
+package com.geekbrains.model;
+
+public class Author {
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    private int id;
+    private String firstName;
+}
