+++ src/main/java/com/geekbrains/model/Product.java
@@ -1,0 +1,44 @@
+package com.geekbrains.model;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class Product {
+    private int id;
+    private String title;
+    private double cost;
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public void setTitle(String title) {
+        this.title = title;
+    }
+
+    public void setCost(double cost) {
+        this.cost = cost;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public String getTitle() {
+        return title;
+    }
+
+    public double getCost() {
+        return cost;
+    }
+
+    @Override
+    public String toString() {
+        return "Product{" +
+                "id=" + id +
+                ", title='" + title + '\'' +
+                ", cost=" + cost +
+                '}';
+    }
+}
+
