+++ src/main/webapp/WEB-INF/templates/add-prod.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" lang="ru"> <head>
+    <title>Add new product</title> </head>
+<body>
+<form th:action="@{'/add_product/form'}" th:object="${product}" method="post">
+    Id: <input type="text" th:field="*{id}"/> <br>
+    Title: <input type="text" th:field="*{title}"/> <br>
+    Cost: <input type="number" th:field="*{cost}"/> <br>
+    <button type="submit">Save</button>
+</form> </body>
+</html>
