+++ src/main/webapp/WEB-INF/templates/repos.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org">
+<head>
+  <title>Product Repository</title>
+</head>
+<body>
+<h1>All Products</h1>
+<p th:each="product : ${repos}" th:text="${product}" />
+</body>
+</html>
